refactor(ShowObsOnPath): MapC에서 길찾기 경로 내 장애물 표시 기능 분리 및 범례 지도에 표시
- 경로 내 장애물 표시 버튼 클릭 시 범례대로 지도에 경로 스타일이 변함
- 계단, 경사로, 비포장도로 클릭은 되나 팝업은 아직 복구 못함
- 경로 내 장애물 표시 버튼 클릭 후 나타난 장애물 아이콘 클릭 시 뜨는 팝업에 x버튼이 모두 적용되지 않는 버그가 있음
- MapC에서 layerstate 부분과 경로 및 마커 표시하는 부분도 분리하려 했으나 경로가 안 뜨는 버그가 생겨서 일단은 그대로 합쳐둠
