+++ src/components/PopupC.js
@@ -6,7 +6,7 @@ import { click } from 'ol/events/condition';
 import {showMarkerStyle} from './MarkerStyle'
 import './MapC.css';
 
-const setPopupSelect = (layer, map) => {
+export const setPopupSelect = (layer, map) => {
     const select = new Select({
         condition: click,
         layers: [layer],
